massive improvements to copypasta (see description)
Before when a copypasta would have an image url and the total length of the copypasta was less than or equal to 1024 it would just be dumped in the MessageEmbed. I have changed this to if the URL starts with http or https and ends with webp, png, jpg or gif it is set as the MessageEmbed image and the URL is snipped out of the copypasta so it's also very cleanly looking.

Before: https://favna.s-ul.eu/YgiQXyh9.png
After: https://favna.s-ul.eu/Lp7wQDaL.png