Major rewrite and new command (See details)
Introduced the util.js which has a lot of common functions so they do not need to be redeclared every time they are used across the command files. This also brings with it a configurable option "timeformat" which can be used to configure whether you want your timestamps (where applicable) to be shown in 24 or 12 hour time format (default is 24).